ai-engine/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
requests
langgraph
faiss-cpu
redis
websockets
